requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.2.1
django-filter==2.2.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-xml==1.4.0
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.9
inflection==0.4.0
itypes==1.1.0
Jinja2==2.11.1
Markdown==3.2.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.3
pathspec==0.8.0
psycopg2==2.8.5
pyparsing==2.4.7
python-decouple==3.3
python-dotenv==0.12.0
pytz==2019.3
regex==2020.4.4
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
whitenoise==5.0.1
